Move admin devices pairing api
